pollsystem/requirements.txt
asgiref==3.11.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
